requirements.txt
requests
urllib3
pycryptodome
m3u8